pyproject.toml
[project]
name = "n8n-executive-reporting-pipeline"
version = "0.1.0"
description = "Automated executive reporting pipeline: n8n + Claude API + Quarto + Slack + Gmail"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "anthropic>=0.40",
    "httpx>=0.27",
    "pydantic>=2.8",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
